-Add support for #includes -Allow scanf calls to struct members -Add scale down project example -Add fake_libc

+++ src/PreProcessing/Examples/example1.c
@@ -1,3 +1,5 @@
+#include<stdlib.h>
+
 int foo(){
     int i = 0;
     return i;
@@ -7,7 +9,7 @@ int GLOBAL = 0;
 
 int main() {
     
-    int i;
+    size_t i;
     long l;
     char array[5];
